Revert "exclude test from race" (#15014)
